+++ lib/pack_core/global_states/ephemeral/level_players_bloc.g.dart
@@ -13,6 +13,8 @@ _$_LiveLevelPlayersBlocState _$$_LiveLevelPlayersBlocStateFromJson(
           .map((e) => PlayerProfileModel.fromJson(e as Map<String, dynamic>))
           .toList(),
       currentPlayerId: json['currentPlayerId'] as String,
+      playerCharacter: PlayerCharacterModel.fromJson(
+          json['playerCharacter'] as Map<String, dynamic>),
     );
 
 Map<String, dynamic> _$$_LiveLevelPlayersBlocStateToJson(
@@ -20,4 +22,5 @@ Map<String, dynamic> _$$_LiveLevelPlayersBlocStateToJson(
     <String, dynamic>{
       'players': instance.players.map((e) => e.toJson()).toList(),
       'currentPlayerId': instance.currentPlayerId,
+      'playerCharacter': instance.playerCharacter.toJson(),
     };