+++ packages/wbw_core/lib/src/models/models.dart
@@ -7,6 +7,7 @@ part 'current_word_model.dart';
 part 'fuel_model.dart';
 part 'fuel_storage_model.dart';
 part 'game_model.dart';
+part 'level_characters_model.dart';
 part 'level_model.dart';
 part 'level_players_model.dart';
 part 'models.freezed.dart';